Move TestComponent to other file
